ai-dungeon-story-generator/requirements.txt
streamlit>=1.25.0
transformers>=4.34.0
torch>=2.0.1
accelerate>=0.20.3
pyyaml
python-dotenv
huggingface-hub>=0.15.1
regex
tqdm